requirements.txt
streamlit~=1.44.0
google-generativeai
python-dotenv~=1.1.0
langchain
langchain_google_genai
PyPDF2
faiss-cpu
protobuf~=5.29.4
langchain-community
nltk